Address round-6 review: observable isolation test, re-materializable per-PID root, stale-PID wipe
- isolation.test.js no longer calls setupTestDBPath() inside the test body,
  so its assertions observe the state mocha.init.js established instead of
  repairing it first (the headline test could not previously fail on the
  regression it exists to catch).
- Factor the per-PID root/config synthesis into unitTests/perPidRoot.js and
  call it from mocha.init.js, setupTestDBPath(), and ensureSystemTables():
  a mid-run tearDownMockDB() removes the whole root including
  harper-config.yaml, and a later re-seed died in updateConfigValue's
  readFileSync instead of re-materializing it.
- mocha.init.js wipes a stale per-PID dir before creating it (a recycled PID
  must not inherit a killed run's config, keys, or seeded system database),
  guarded on isMainThread so worker threads re-running the preload under the
  same PID cannot delete the live root.
- mTLS tests assert cert.private_key_name before joining it into a key path;
  the seeded cert-record rename derives the name from testKeyPath instead of
  a second hardcoded copy.
- Trim comments that restated adjacent code.
