+++ unitTests/apiTests/mqtt-test.mjs
@@ -608,6 +608,7 @@ describe('test MQTT connections and commands', function () {
 		}
 		cert ??= fallbackCert;
 		assert.ok(cert, 'expected a non-authority certificate record in system.hdb_certificate');
+		assert.ok(cert.private_key_name, 'expected the certificate record to name its private key file');
 		// resolve the key the same way the server does: by the cert record's key name under <root>/keys
 		const private_key_path = join(env_get('rootPath'), 'keys', cert.private_key_name);
 		let client = await connectAsync(`mqtts://${testHost}:8884`, {
@@ -681,7 +682,7 @@ describe('test MQTT connections and commands', function () {
 			}
 			cert ??= fallbackCert;
 			assert.ok(cert, 'expected a non-authority certificate record in system.hdb_certificate');
-			// resolve the key the same way the server does: by the cert record's key name under <root>/keys
+			assert.ok(cert.private_key_name, 'expected the certificate record to name its private key file');
 			const private_key_path = join(env_get('rootPath'), 'keys', cert.private_key_name);
 			let bad_client = await connectAsync(`wss://${testHost}:8885`, {
 				reconnectPeriod: 0,
